Issue #377 - use the VMUSB drivers in place rather than as copies
Import the caenchain directly from vmusb/daqconfig rather than a
maintaince headache prone local copy.
