+++ main/usb/vmusb/daqconfig/CMADCChain.h
@@ -13,36 +13,21 @@
 	     Michigan State University
 	     East Lansing, MI 48824-1321
 */
+/**
+ * @file CMADCChain.h 
+ * @brief Header for the Mesytec chain manager for Mesytec devices that can CBLT.
+ * @note WHen compiled with mvlcgenerate MVLC_GENERATOR is defined making this
+ * work in both VMUSB and MVLC.
+ */
 #ifndef __CMADCCHAIN_H
 #define __CMADCCHAIN_H
-
-#ifndef __CREADOUTHARDWARE_H
-#include "CReadoutHardware.h"
-#endif
-
-#ifndef __CMESYTECBASE_H
-#include "CMesytecBase.h"
-#endif
-
-#ifndef __CRT_STDINT_H
+#include <CReadoutHardware.h>
+#include <CMesytecBase.h>
 #include <stdint.h>
-#ifndef __CRT_STDINT_H
-#define __CRT_STDINT_H
-#endif
-#endif
-
-#ifndef __STL_STRING
 #include <string>
-#ifndef __STL_STRING
-#define __STL_STRING
-#endif 
-#endif
-
-#ifndef __STL_LIST
 #include <list>
-#ifndef __STL_LIST
-#define __STL_LIST
-#endif
+#ifdef MVLC_GENERATOR
+#include <DeviceCommand.h>
 #endif
 
 
@@ -50,6 +35,13 @@ class CReadouModule;
 class CVMUSB;
 class CVMUSBReadoutList;
 class CMADC32;
+
+#ifdef MVLC_GENERATOR
+class TCLConfigParser;
+namespace XXUSB { 
+  class CConfigurableObject;
+}
+#endif
 
 /*!
    This class implements a chain of MADC32's to be read out via CBLT.
@@ -83,13 +75,24 @@ private:
   // Per object data:
 private:
   ChainList         m_Chain;	// List of modules first must be leftmost, last must be right most.
+#ifdef MVLC_GENERATOR
+  XXUSB::CConfigurableObject*   m_pConfig;  // My configuration database.
+  TCLConfigParser* m_pParser;
+#else
   CReadoutModule*   m_pConfig;  // My configuration database.
-
+#endif
 
 public:
+#ifdef MVLC_GENERATOR
+  CMADCChain(TCLConfigParser* parser);
+#else
   CMADCChain();
-  CMADCChain(const CMADCChain& rhs);
+#endif
   virtual ~CMADCChain();
+#ifndef MVLC_GENERATOR
+private:
+#endif
+  CMADCChain(const CMADCChain& rhs);
   CMADCChain& operator=(const CMADCChain& rhs);
 
   // Disallowed canonicals:
@@ -103,11 +106,16 @@ public:
   // overridable oeprations on constructed objects:
 
 public:
+#ifdef MVLC_GENERATOR
+  virtual void onAttach(XXUSB::CConfigurableObject& configuration);
+#else
   virtual void onAttach(CReadoutModule& configuration);
+#endif
   virtual void Initialize(CVMUSB& controller);
   virtual void addReadoutList(CVMUSBReadoutList& list);
+#ifndef MVLC_GENERATOR
   virtual CReadoutHardware* clone() const;
-
+#endif
   // utilities:
 
 private:
@@ -117,4 +125,21 @@ private:
   static bool moduleChecker(std::string name, std::string value, void* arg);
   
 };
+
+#ifdef MVLC_GENERATOR
+/**
+ * @class CMADCChainCommand
+ * 
+ *   Command/generator that defines the "madcchain" command to create/config chains.
+ */
+
+class CMADCChainCommand : public DeviceCommand {
+  TCLConfigParser* m_parser;                   // Needed to pass to chains.
+public:
+  CMADCChainCommand(CTCLInterpreter& interp, TCLConfigParser& parser) ;
+  virtual ~CMADCChainCommand();
+protected:
+  CReadoutModule* createDevice(std::string name);
+};
+#endif
 #endif
